La til @vtfk/logger
