+++ lib/invoke-ps-file.js
@@ -1,3 +1,4 @@
+const { logger } = require('@vtfk/logger')
 const { exec } = require('child_process')
 
 const { MAX_BUFFER } = require('../config')
@@ -17,18 +18,18 @@ const invoke = (filePath, caller, args) => {
 
     // set encoding directly in the console: "cmd.exe /c chcp 65001>nul &&"
     const cmd = `cmd.exe /c chcp 65001>nul && powershell.exe -NoLogo -ExecutionPolicy ByPass -Command "${filePath}"${args ? ` ${parseArgs(args)}` : ''}`
-    console.log(`${caller} :: exec starting :: ${new Date().toISOString()} :: exec cmd:`, cmd)
+    logger('info', `${caller} :: exec starting :: ${new Date().toISOString()} :: exec cmd:`, cmd)
     const proc = exec(cmd, { maxBuffer: Number.parseInt(MAX_BUFFER) }, (error, stdout, stderr) => {
       if (stderr !== '') {
-        console.log(`${caller} :: exec stderr :: ${new Date().toISOString()} :: ${proc.pid}`)
+        logger('error', `${caller} :: exec stderr :: ${new Date().toISOString()} :: ${proc.pid}`)
         return reject(new Error(sanitizeError(filePath, stderr)))
       }
       if (error !== null) {
-        console.log(`${caller} :: exec error :: ${new Date().toISOString()} :: ${proc.pid}`)
+        logger('error', `${caller} :: exec error :: ${new Date().toISOString()} :: ${proc.pid}`)
         return reject(new Error(sanitizeError(filePath, error)))
       }
 
-      console.log(`${caller} :: exec finished :: ${new Date().toISOString()} :: ${proc.pid}`)
+      logger('info', `${caller} :: exec finished :: ${new Date().toISOString()} :: ${proc.pid}`)
       return resolve(stdout)
     })
   })
