+++ lib/is-valid-json.js
@@ -1,0 +1,9 @@
+const isValidJSON = str => {
+  try {
+    return JSON.parse(str)
+  } catch (error) {
+    return false
+  }
+}
+
+module.exports = isValidJSON
