+++ lib/sanitize-error.js
@@ -1,6 +1,8 @@
+const { logger } = require('@vtfk/logger')
+
 module.exports = (filePath, error) => {
   if (error.includes(`${filePath} : `)) {
-    console.log('sanitize-error: Removing "filePath : " and substringing error')
+    logger('info', 'sanitize-error: Removing "filePath : " and substringing error')
     error = error.replace(`${filePath} : `, '')
 
     return error.substring(0, error.indexOf('At line:')).trim()
