+++ package.json
@@ -12,6 +12,7 @@
   "author": "Rune Moskvil Lyngås <runely85@gmail.com>",
   "license": "ISC",
   "dependencies": {
+    "@vtfk/logger": "2.4.0",
     "child_process": "1.0.2",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
